X- Resolver problema de conversión de float a int/decimal
git-svn-id: file:///d/A__Repository/AnTScript/trunk@15 85c3cb87-3462-2548-830f-76f4d12b9134

+++ AnTScript/_ObjTest.cs
@@ -7,7 +7,7 @@ namespace AnTScript
 {
     public class _Node
     {
-        public int IdNota { get; set; }
+        public decimal IdNota { get; set; }
         public string Asunto { get; set; }
         public string Nota { get; set; }
 
@@ -30,7 +30,6 @@ namespace AnTScript
         public string Descripcion { get; set; }
         public string Nota { get; set; }
         
-
         public _Folder()
         {
             IdCarpeta = 1;
